# Cgroup2: Reduce allocations for `manager.Stat`