@dcantah 2023-04-07
Working towards https://github.com/containerd/cgroups/issues/280. There's still some spots we can fix, as the majority of allocs are occurring in bufio.Scan in readKVStatsFile