@dcantah 2023-04-07
> I suggest making `readHugeTlbStats` changes related to readdirnames as a separate commit.

That's fair, they were all in the pursuit of improving Stat so I felt either way. I originally had the readSingleFile and uint improvements in a separate commit and squished after finding the other areas. I'll separate them though