@kolyshkin 2023-04-07
By looking at the code, I'd rather have it refactored first, then optimize. At least some parts that are crying to be removed, such as readSingleFile -- PTAL https://github.com/containerd/cgroups/pull/283