@dcantah 2023-04-24
Okay, updated to include the runc trick @kolyshkin, great suggestion. We're down to 199 now.