@kolyshkin 2023-04-27
> All in all on Go 1.19.4 this drops allocations in Stat from 322 to 199, or 47%.

Oh, and if you can use `benchstat` for such calculations that'd be terrific.