@dcantah 2023-04-28
> Oh, and if you can use `benchstat` for such calculations that'd be terrific.

No wonder you recommended, I fat fingered 47 from 38.. No more programming past 12 🐸